Rework the global scripts to work in script mode
