+++ src/builtins/internal_js_modules/node/async_hooks.js
@@ -253,10 +253,7 @@ export class AsyncLocalStorage {
         return AsyncResource.bind(fn);
     }
     static snapshot() {
-        return AsyncLocalStorage.bind((
-            cb,
-            ...args
-        ) => cb(...args));
+        return AsyncLocalStorage.bind((cb, ...args) => cb(...args));
     }
 }
 export function executionAsyncId() {
